+++ CHANGELOG.md
@@ -1,12 +1,16 @@
-# websitetoolboxforum Changelog
+# website toolbox community Changelog
 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.4.0 - 2023-02-14
+### Added
+- Added support for CraftCMS 4.
+ 
 ## 1.3.0 - 2021-12-27
 ### Added
-- Website Toolbox forum Settings issue resolved.  
+- Website Toolbox Community Settings issue resolved.  
 
 ## 1.0.0 - 2020-06-30
 ### Added